CI fix: add missing includes (heap.h, spinlock.h, kptr_restrict.h), fix narrowing conversions (uint64→uint32), suppress unused variable
Errors identified from GitHub Actions CI runs:
- nvme.c: kmalloc/kfree undeclared → add #include "heap.h"
- net_tcp.c: spinlock_t unknown → add #include "spinlock.h"
- procfs.c: kptr_restrict_check undeclared → add #include "kptr_restrict.h"
- vfs.c: uint64_t→uint32_t at 9 call sites (existing_size, new_size, old_size, bytes_to_blocks, dcache_add, read/write size params)
- elf.c: size_t→int casts at 4 strlen() sites, commented unused user_stack_guard
- idt.c: uint64_t→uint32_t cast on offset_high assignment

+++ src/kernel/idt.c
@@ -75,7 +75,7 @@ static void idt_set_gate(uint8_t num, uint64_t handler, uint16_t sel, uint8_t ty
     idt[num].ist         = 0;
     idt[num].type_attr   = type_attr;
     idt[num].offset_mid  = (handler >> 16) & 0xFFFF;
-    idt[num].offset_high = (handler >> 32) & 0xFFFFFFFF;
+    idt[num].offset_high = (uint32_t)((handler >> 32) & 0xFFFFFFFF);
     idt[num].reserved    = 0;
 }
